Move metadata stores to zenml.metadata
