+++ zenml/metadata/__init__.py
@@ -12,7 +12,7 @@
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
 
-from zenml.core.metadata.metadata_wrapper import ZenMLMetadataStore
-from zenml.core.metadata.mysql_metadata_wrapper import MySQLMetadataStore
-from zenml.core.metadata.mock_metadata_wrapper import MockMetadataStore
-from zenml.core.metadata.sqlite_metadata_wrapper import SQLiteMetadataStore
+from zenml.metadata.metadata_wrapper import ZenMLMetadataStore
+from zenml.metadata.mock_metadata_wrapper import MockMetadataStore
+from zenml.metadata.mysql_metadata_wrapper import MySQLMetadataStore
+from zenml.metadata.sqlite_metadata_wrapper import SQLiteMetadataStore
