+++ zenml/metadata/sqlite_metadata_wrapper.py
@@ -16,7 +16,7 @@ from typing import Text
 
 from tfx.orchestration import metadata
 
-from zenml.core.metadata import ZenMLMetadataStore
+from zenml.metadata import ZenMLMetadataStore
 from zenml.utils import path_utils
 from zenml.utils.enums import MLMetadataTypes
